assistant-mvp/requirements.txt
requests==2.31.0
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.6.4
python-telegram-bot==21.0.1
